feat: component 패키지 분리(#30)
