+++ feature/quiz/src/main/java/kr/boostcamp_2024/course/quiz/component/QuizSolveTimeSlider.kt
@@ -1,0 +1,58 @@
+package kr.boostcamp_2024.course.quiz.component
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.defaultMinSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Slider
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableFloatStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun QuizSolveTimeSlider(
+    onValueChange: (String) -> Unit
+) {
+    var sliderPosition by remember { mutableFloatStateOf(0f) }
+
+    Slider(
+        modifier = Modifier.fillMaxWidth(),
+        value = sliderPosition,
+        onValueChange = { sliderPosition = it },
+        steps = 8,
+        valueRange = 10f..100f,
+        thumb = {
+            Box(
+                modifier = Modifier
+                    .defaultMinSize(minWidth = 40.dp)
+                    .clip(MaterialTheme.shapes.large)
+                    .background(MaterialTheme.colorScheme.primary),
+                contentAlignment = Alignment.Center
+            ) {
+                Text(
+                    text = (sliderPosition.toInt()).toString(),
+                    color = MaterialTheme.colorScheme.onPrimary,
+                    textAlign = TextAlign.Center
+                )
+            }
+        }
+    )
+}
+
+@Preview(showBackground = true)
+@Composable
+fun QuizSolveTimeSliderPreview() {
+    QuizSolveTimeSlider(onValueChange = {})
+}
